TST: GH26807 remove classes and subdirectory, move error tests from test_read into test_errors
